ui/requirements.txt
fastapi>=0.100.0
uvicorn>=0.20.0
python-dotenv>=1.0.0
pyzmq>=25.0.0
openai>=1.0.0
